requirements.txt
fastapi==0.135.2
google-genai==1.11.0
pandas==2.2.3
pydantic==2.12.5
python-dotenv==1.1.1
uvicorn==0.42.0
yfinance==1.2.0
